feat: `play` command changes; `resume` command
play: set `query` as required and remove `resume playback` function
resume: add `resume` command to resume playback
